+++ OneBlink.SDK/models/FormApprovalFormPrefillMapping.cs
@@ -1,0 +1,14 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormApprovalFormPrefillMapping
+    {
+        public FormApprovalPrefillElementPath sourcePath
+        {
+            get; set;
+        }
+        public FormApprovalPrefillElementPath approvalPath
+        {
+            get; set;
+        }
+    }
+}